new UI for members page
