+++ next.config.js
@@ -1,7 +1,7 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
     images: {
-        domains: ["user-images.githubusercontent.com", "i.ibb.co", "images.unsplash.com", "media.geeksforgeeks.org"],
+        domains: ["user-images.githubusercontent.com", "i.ibb.co", "images.unsplash.com", "media.geeksforgeeks.org","drive.google.com"],
     },
 };
