+++ package.json
@@ -16,6 +16,7 @@
         "@prisma/client": "^4.12.0",
         "axios": "^1.3.2",
         "bcrypt": "^5.1.0",
+        "codex-website": "file:",
         "cors": "^2.8.5",
         "discord.js": "^14.7.1",
         "dotenv": "^16.0.3",
