+++ src/components/proud_alumni.tsx
@@ -91,62 +91,12 @@ const Proud_alumni = () => {
             github: "https://github.com/AriefB18",
             linkdin: "https://www.linkedin.com/in/arief-basha-a3b558236",
         },
-        {
-            image: "https://user-images.githubusercontent.com/105535366/222389442-4b7ec2d0-bfbb-4106-b686-9aa735d6f1bc.png",
-            name: "Bharath",
-            role: "Logistics Officer",
-            description:
-                "Manages all the Logistical requirements and Clerical work along with coordinating and optimizing the movement of goods and resources that's necessary",
-            insta: "https://instagram.com/_.b.h.a.r.a.t.h.__.y.a.d.a.v._?igshid=YmMyMTA2M2Y=",
-            github: "#",
-            linkdin: "#",
-        },
-        {
-            image: "https://user-images.githubusercontent.com/105535366/222389447-89270e64-8ba1-41d8-b982-5999e056fcea.png",
-            name: "Tamana",
-            role: "Public Relations Officer",
-            description:
-                "Craft compelling stories and Be the voice of our CODEX and build meaningful connections.",
-            insta: "https://instagram.com/truly_tamanna?igshid=YmMyMTA2M2Y=",
-            github: "https://www.linkedin.com/in/tamannashaw",
-            linkdin: "#",
-        },
-        {
-            image: "https://user-images.githubusercontent.com/105535366/222389455-99366e0b-67f7-45aa-bf29-c38549b270f7.png",
-            name: "Suhit",
-            role: "Visual Arts Officer",
-            description:
-                "The one responsible for leading and to create visually stunning and technically precise art assets that help in marketing and promotional aspect of CODEX ",
-            insta: "https://instagram.com/suhit_eswar_?igshid=YmMyMTA2M2Y=",
-            github: "https://github.com/suhit123",
-            linkdin: "#",
-        },
-        {
-            image: "https://user-images.githubusercontent.com/105535366/222389435-77dd1d02-3543-4d69-9b4f-c3dfeedeffb6.png",
-            name: "Hilux",
-            role: "Front End - Lead",
-            description:
-                "Incharge of creating intuitive and visually stunning user experiences that immerse the user into the project",
-            insta: "https://www.instagram.com/fokouhilux/",
-            github: "https://github.com/hiluxfnh",
-            linkdin: "https://www.linkedin.com/in/fokou/",
-        },
-        {
-            image: "https://user-images.githubusercontent.com/141537855/258637418-ecd74f81-067a-47af-a737-b9aff33ce430.png",
-            name: "Jasir",
-            role: "Non Tech Admin",
-            description:
-            "Efficiently manage administrative tasks and operations in a non-technical capacity using Codex.",
-            insta: "https://instagram.com/jasirmd33?utm_source=qr&igshid=ZDc4ODBmNjlmNQ%3D%3D",
-            github: "https://github.com/Jasirmd",
-            linkdin: "https://www.linkedin.com/in/jasir-mohammed-4149ba246",
-        },
     ];
     return (
         <div className={styles.proud_alumni_codex}>
             <div className={styles.team_codex_div}>
                 <h1 className={styles.team_codex_heading}>
-                    Proud Alumni of <span>CODEX</span>
+                    <span>v.1.0</span>
                 </h1>
                 <div className={styles.member_blocks_team_codex}>
                     {data.map((item, index) => (
